feat: add PreviewTab component and integrate it into DesktopEditor and editor page
